+++ src/BloomBrowserUI/bookEdit/js/bloomImages.ts
@@ -21,7 +21,6 @@ import { updateCanvasElementClass } from "../toolbox/canvas/canvasElementDomUtil
 import { farthest } from "../../utils/elementUtils";
 import { EditableDivUtils } from "./editableDivUtils";
 import { playingBloomGame } from "../toolbox/games/DragActivityTabControl";
-import { kPlaybackOrderContainerClass } from "./talkingBookMarkupConstants";
 import { getWorkspaceBundleExports } from "./workspaceFrames";
 import {
     changeImage,
@@ -651,8 +650,9 @@ export function handleMouseEnterBloomCanvas(bloomCanvas: HTMLElement): void {
     SetImageTooltip(bloomCanvas);
 
     if (
-        bloomCanvas.getElementsByClassName(kPlaybackOrderContainerClass)
-            .length > 0
+        bloomCanvas.getElementsByClassName(
+            "bloom-playbackOrderControlsContainer",
+        ).length > 0
     ) {
         return; // Playback order controls are active, deactivate bloom-canvas stuff.
     }
