--- src/BloomBrowserUI/bookEdit/js/talkingBookMarkupConstants.ts
@@ -1,8 +1,0 @@
-// taken out of audioRecording.ts to avoid the need for other
-// files to import that big file just to use a little bit of
-// code.
-
-export const kPlaybackOrderContainerClass: string =
-    "bloom-playbackOrderControlsContainer";
-
-export const kAudioCurrent = "ui-audioCurrent";
